Move showup_core package to project root
