+++ showup_core/batch_persistence.py
