+++ showup_core/data/logs/workflow2/core.log
@@ -15,3 +15,10 @@
 2025-07-13 14:08:09,485 - core - INFO - Initializing ShowupSquared core module
 2025-07-13 14:24:50,210 - core - INFO - Initializing ShowupSquared core module
 2025-07-13 15:58:43,266 - core - INFO - Initializing ShowupSquared core module
+2025-07-13 14:30:34,685 - core - INFO - Initializing ShowupSquared core module
+2025-07-13 14:30:37,025 - core - INFO - Initializing ShowupSquared core module
+2025-07-13 14:30:39,767 - core - INFO - Initializing ShowupSquared core module
+2025-07-13 14:30:45,122 - core - INFO - Initializing ShowupSquared core module
+2025-07-13 14:30:49,892 - core - INFO - Initializing ShowupSquared core module
+2025-07-13 14:30:53,522 - core - INFO - Initializing ShowupSquared core module
+2025-07-13 14:30:56,775 - core - INFO - Initializing ShowupSquared core module
