+++ showup_core/editor.py
